+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Remove `vue` from `disableLanguages` because Vue is supported in Prettier 1.15 version.
 - Move `eslint`, `tslint` and `stylelint` integrations to separate file `integrations.ts` (out from `utils.ts`) and `require()` (load) dependency only when it's needed and cache it in-memory.
 - Remove local prettier support.
+- Remove unnecessary settings from `test.code-workspace` file.
 
 ## 2.0.0 (August 16, 2019)